Fix invalid solution bug found on 8-1
Now i'm pretty sure the solver works correctly. I'm not totally
convinced that it's laid out in the most sensible way, because the
method "_is_solved()" does so many things: returns False if we're
not down to 1 candidate in each cell, raises a NoSolutionError if
we have a proposed solution that is invalid, and finally, if we
have a valid solution, it saves it and then returns True. That's a
lot for one method. Some restructuring could be warranted.